Big update here: timetravel mode no longer uses an extra shell child to monitor command execution. i.e. the minimal number of processes is used
In timetravel mode recursive_execute_command and
close_command_exec_resources use NONBLOCKING waits. If the executing
command has finished at the time of the check, the next command is
started (if applicable). Otherwise the commands will return. The next
subsequent call on them will check the status again and if it has
finished the next logical step is taken.

Thus as the timetravel() function loops through each command, the
status is periodically checked, updated, and new commands are fired as
appripriate.

@pcrumm this update was originally intended to be a part of the design
problem, however, all test cases pass so I don't see why not to add it
Note: the test runs differently under `make check` and `make dist`
one of the lines from the first test comes out of order. If this occurs
before submission, I'm not sure if we should submit the script that
works with make dist or make check...