Timetrash now takes pipe chains into account when excecuting with a specified limit of parallelization
@pcrumm If no limit is specified I set the number of available
processes for runningt o LONG_MAX. It seems a bit arbitrary, but it was
the simplest way I could think of to allow many commands to run at once
without having to constantly check for no limits. Do you have any
suggestions on improving this, or do you think this arbitrary limit is
permissable here?