Implemented read and write to virtual disk
@pcrumm I have not used any sort of locking in function
osprd_process_request. The osprd_info_t has a spin lock type associated
with it, however, no native locking of its type has been implemented
(besides checking for invalid locks and unlocks). I'm guessing this is
because we are expected to implement disk locking with it.

Do you think it is safe to assume that if control has reached the
processing of a request we have exclusive access to the disk, or need
we double check that we have proper locking?