Fixed bug with processes that get killed, all tests now pass
* If a blocked process receives a signal, the ticket head and tail are
  cleared and all other processes in the queue are woken up and forced
  to requeue (so there are no gaps between pending tickets).

@pcrumm If several processes receive signals, we requeue EVERYTHING
once for each of the processes. If there are a lot of pending jobs and
a lot of signals being intercepted, our module would take a while to
finish getting organized.

The current implementation is the easiest and
simplest way I could come up with to get the job done (and avoid the
headaches associated with maintaining an array of valid tickets or a
linked list or something other).

Thoughts?