Initial implementation of deadlocking
@pcrumm this works on simpler deadlock cases such as one process asking
for the same lock twice, or two processes each requesting the two same
locks. Our test case with the three processes and three locks is NOT
handled properly. I tried changing the delay times, but I couldn't
figure out if that was the issue or whether there is a bug in the code.
Please see if you can shed some light on the issue.