Fixed add_block() and remove_block() to behave more robustly
When a file is first created it will have a single block allocated to
it, but at any time after that it is possible that it may have NO
allocated blocks (ie if change_size(oi, 0) was called on that file).
Thus add_block() and remove_block() will take into account when to
add or free the initial block or ignore it when needed.

@pcrumm This approach creates the possible side effect of truncating a
file and if the disk later fills up it will not be able to grow (as
opposed to the case where a newly created file will always have at
least one block to write to). I don't think this deviates too far from
what may happen on Linux, however...