All download tasks are now done by the child, including starting and completing a download
There is an issue where sometimes file checksums are not properly read
and the client will run into an invalid response from the tracker,
causing it to panic and die. As this occurs randomly, it is MUCH more
preferential for only a child to abort rather than bring down the
entire client and any other valid processes.